Enhance graphviz options
Add 'graphviz-no-branches' to output a history graph with no branch names. This
is basically the original example I started with.

Rename our previous 'graphviz' to 'graphviz-with-branches' to be more explicit,
set the 'graphviz' alias to 'graphviz-with-branches'. This will let folks using
this config switch the default if they'd like to